Add files via upload
This is another Project which is developed by me along with a team of 4 members. Where we used Angular for frontend, Spring Boot for the middle tag and MySQL for the backed. 
Investment Tracking is a full Stack Web Development Project where there are user and admin roles. User can track their Investments made by the Admin. If the user have any quires he can Inquire about the Investment and he can give his feedback according to his opinion.